Fixes #14. Also fixes minor ship movement error.
Before, if an empty ship was told to move, and was then garrisoned, it
would move to where it was told to go before. Now it stays still.
Also, fixed very minor file include issue.
This fix may have generated some issues with obstacle avoidance and
ships. It doesn't appear to have done so, but the possibility remains.