Fixed two sudden bugs.
Bug 1: Crash on startup, because I skipped an argument (attackarea) in
unit.specs. Quickly and easily remedied.

Bug 2: Crash when finished training a unit while the building doing the
training is not selected. Fixed by pseudo-selecting the building and
then reselecting the proper units/buildings/whatever. So kind of a hacky
fix, but it worked.