Closes #21.
Animals can move, be hunted, and reproduce. They have varying health,
and they run when you attack them. There is sometimes perceptible lag
while hunting.
I think this is due to unit::searchresources, but I'm not
sure. It may be wise to write a version of this function that uses a 3
by 3 square instead of a 9 by 9 square, since the animal should be
within 1 square of its old position. Instead, maybe even just search the
four cardinal directions away from the old position. That ought to
usually do it, unless obstacle avoidance was involved. Or maybe its not
that function.